Remove broken `state first` feature (#64)
Since hyperapp/router relies on History API especially "popstate" event,
developers must add an event listener to `window`. In general, calling
`location.subscribe()` acts for the purpose, but the function needs
`location.actions` as an argument. This means that developers must wire
`location` slice to `actions`. Then `location.actions` updates
`state.location` not `window.location` so developers need to wire it to
`state` as well.

Consequently, there is no way to skip wiring the `location` slice to
`state` and `actions`.

Another solusion is to provide `location.subscribeWindowWithoutActions()`
but may be it should be done by another npm package.